requirements.txt
Flask-RESTful==0.3.8
PyJWT>=2.0,<3.0
Flask-SQLAlchemy>=3.0.0
